Implement -p flag for blast
Some major refactoring here. I wrote a private function to get blast binaries out of -p, and did my best to make the main function flexible. It can handle binaries outside of $PATH, but they still need to be named 'blastn', 'blastp', and 'blastdbcmd' for the program to work. Maybe try to fix that in the future.